Add analyze part
Use difflib library to get replace part as hits.
Add limit to the empty web contents.
Add crawlflow to record the crawling flow.

+++ src/google_crawler.py
@@ -11,7 +11,6 @@ import logging
 import sys
 import os
 import glob
-import random
 import json
 from os.path import join
 from time import sleep
@@ -24,6 +23,9 @@ import multiprocessing as mp
 import subprocess
 import jieba, jieba.analyse
 import shutil
+from pymongo import MongoClient
+import datetime
+from difflib import *
 
 class ASRdataFetcher(object):
     """Summary of class here.
@@ -206,30 +208,55 @@ def crawlpage(url):
     except:
         alldata = ''
     return alldata
-def analyze(filename,outputpath,threshold,thisTurnData,input_text_path):
+def analyze(filename,outputpath,threshold,thisTurnData,input_text_path,crawlflow):
+    """Analyze match result; report whether re-crawl or not
+    
+    Open match file and consider the score of match result.
+    Copy processed file and generate some tracking files.
+    
+    Args:
+        filename: asr file name
+        outputpath: the location of match file
+        threshold: besed on this threshold to decide re-crawl or not
+        thisTurnData: the web contents lists from extended function crawlpage output
+        input_text_path: loading folder
+        crawlflow: the dict to store in database
+    Returns:
+        State,crawlflow:
+            ex:
+            'Get paragraph',crawlflow
+            'Crawl Again',crawlflow
+    """
     with open(join(outputpath,'fcr23.ws.re.wav.all2.match'),'r',encoding='utf8') as f:
         next(f)
         data = f.readline().strip()
         assert len(data.split('\t')) == 17
         score = data.split('\t')[16]
+        filepath = data.split('\t')[0]
         #print('crawl '+str(len(thisTurnData))+' pages, spend '+
         #'crawlPagetime:{}\ttranfPinYintime:{}\tmatchFunctiontime:{}'
         #.format(crawlPagetime,tranfPinYintime,matchFunctiontime))
         if float(score) > float(threshold):
-            print(data.split('\t')[10])    
+            paragraph = data.split('\t')[10]
+            crawlflow['filepath'] = filepath
+            crawlflow['score'] = score
+            crawlflow['paragraph'] = data.split('\t')[10]
+            print(paragraph)    
             #write ASR result and web data
             with open('final','a',encoding='utf8') as f:
                 thispath = [path for path in input_text_path if filename in path][0]
                 fetch = ASRdataFetcher()
-                f.write(''.join(fetch.get(thispath,0))+'\t'+data.split('\t')[10]+'\n')
+                crawlflow['oriASRresult'] =''.join(fetch.get(thispath,0))
+                f.write(crawlflow['oriASRresult']+'\t'+paragraph+'\n')
             [shutil.copy(fname, join(outputpath,'finish')) for fname in glob.glob(join(outputpath,('*.txt')))]    
             shutil.copy(join(outputpath,'fcr23.ws.re.wav.all2.match'), join(join(outputpath,'finish'),filename+'match'))
-            return 'Get paragraph'
+            return 'Get paragraph',crawlflow
         else:
             
             print('not found')
             [shutil.copy(filename, join(outputpath,'finish')) for filename in glob.glob(join(outputpath,('*.txt')))]   
-            return 'Crawl Again'
+            return 'Crawl Again',crawlflow
+
 logger = logging.getLogger('google_crawler')
 logger.setLevel(logging.DEBUG)
 formatter=logging.Formatter('%(asctime)s - %(filename)s[line:%(lineno)d] - %(levelname)s: %(message)s')
@@ -241,40 +268,48 @@ logger.addHandler(fh)
 def main():
     # varibales
     input_text_folder = join('..','input_ASR_results')
-    #googletopN = '10'
-    #outputpath = join('..','crawlresult')
+    conn = MongoClient('localhost',27017)
+    db = conn.googlecrawlstream
+    timestamp = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+    
+    #load config
     with open('config','r',encoding='utf8') as f:
         config = json.loads(f.readlines()[0].strip())
     outputpath = config['outputpath']
     Nasgoogle_crawl_dir = config['Nasgoogle_crawl_dir']
     ASR_result = config['ASR_result']
     bashfilepath =config['bashfilepath']
-
+    input_text_folder = config['input_text_folder']
     # load from input text path
     input_text_path = [join(input_text_folder,os.path.basename(x)) for x in glob.glob(join(input_text_folder,('*'))) 
                     if '.cm' in x and '.cm2' not in x and '.syl' not in x]
     #print(input_text_path)
-    
+    input_text_path = sorted(input_text_path, key=functools.cmp_to_key(myCompare))
     filetoUrls = {}
     matchfile_pre = 'fcr23.ws.re.wav.all2'
     matchfile_tmp = 'fcr23.ws.re.wav.all2.res'
     matchfile_result = 'fcr23.ws.re.wav.all2.match'
     for eachTarget in [reconstruct_search_words(eachpath,0.845) for eachpath in input_text_path]:
         for filename, keywordlist in eachTarget.items():
+            crawlflow = {}
             # get web urls from google each 15 seconds
+            crawlflow['filename'] = filename
             logger.info('Start: '+filename)
             n_segment_urls = {}                 # is there a repetition in urls
             alldata = []
             print(filename,keywordlist)
-            #if not filename == 'A0000558':
-            #    continue
+            if not filename == 'A0000048':
+               continue
             thisTurnData = []
+            crawlflow['keywordlist'] = keywordlist
             for keyword in keywordlist:
+                crawlflow['keyword'] = keyword
                 [os.remove(filename) for filename in glob.glob(join(outputpath,('fcr23.ws.re.wav*')))]
                 [os.remove(filename) for filename in glob.glob(join(outputpath,('*.txt')))]
                 [os.remove(filename) for filename in glob.glob(join(outputpath,('*.line')))]
                 tFirstStart = time.time()
                 webUrls = get_web_url(keyword)  # crawl google
+                crawlflow['webUrls'] = webUrls
                 for url in webUrls:
                     if url in n_segment_urls:
                         n_segment_urls[url] += 1
@@ -283,11 +318,15 @@ def main():
                         n_segment_urls[url] = 1
                 pool = mp.Pool()
                 thisTurnData = pool.map(crawlpage,webUrls)
+                
                 alldata.extend(thisTurnData)
                 pool.close()
                 pool.join()
                 crawlPagetime = str(int(time.time()-tFirstStart))
-                
+                crawlflow['crawlPagetime'] = crawlPagetime
+                thisTurnData = [data for data in thisTurnData if len(''.join(data)) < 30000 and not data == '']
+                if not thisTurnData:
+                    break
                 # write down those data from web page
                 for data in thisTurnData:
                     webcontent = ''.join(data)
@@ -298,6 +337,7 @@ def main():
                 tStart = time.time()
                 res = rq.get(bashfilepath+'?text={}&asr={}'.format(Nasgoogle_crawl_dir,ASR_result))
                 tranfPinYintime = str(int(time.time()-tStart))
+                crawlflow['tranfPinYintime'] = tranfPinYintime
                 tStart = time.time()
                 # use match method to find paragraph
                 
@@ -308,8 +348,18 @@ def main():
                     join(outputpath,matchfile_result)],cwd="Match/wav_matched/",stdout=subprocess.PIPE,shell=True)
                 p2.wait()
                 matchFunctiontime = str(int(time.time()-tStart))
-                # read match file and decide to query this file or not
-                if analyze(filename, outputpath, 0.9, thisTurnData, input_text_path) == 'Get paragraph':
+                crawlflow['matchFunctiontime'] = matchFunctiontime
+                # Analyze - read match file and decide to query this file or not
+                if analyze(filename, outputpath, 0.9, thisTurnData, input_text_path,crawlflow)[0] == 'Get paragraph':
+                    # from crawlflow['oriASRresult'] to compare with crawlflow['paragraph']
+                    crawl_compare_match = SequenceMatcher(None, crawlflow['oriASRresult'], crawlflow['paragraph']).get_matching_blocks()
+                    same_sents = [crawlflow['oriASRresult'][m[0]:m[0]+m[2]] for m in crawl_compare_match]
+                    same_sents = [sentence for sentence in same_sents if len(sentence) > 1]
+                    opc1=SequenceMatcher(None, crawlflow['oriASRresult'], crawlflow['paragraph']).get_opcodes()    
+                    hints = [crawlflow['paragraph'][j1:j2] for tag, i1, i2, j1, j2 in  opc1 if tag == 'replace']
+                    hints.extend(same_sents)
+                    crawlflow['hints'] = hints
+                    db[timestamp].insert_one(crawlflow)
                     break
                 else:
                     thisTurnData = []
